+++ CHANGELOG.md
@@ -6,4 +6,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased - 0.0.1]
 
 ### Added
-* [2021-02-05] [PR #1](https://github.com/xenit-eu/docker-haproxy-alfresco/pull/1) Initial version.
+* [2021-04-22] [PR #2](https://github.com/xenit-eu/docker-haproxy-alfresco/pull/2) Added support for ooi service backend.
+* [2021-02-05] [PR #1](https://github.com/xenit-eu/docker-haproxy-alfresco/pull/1) Initial version.